Some basic initial updates
Create very bare bones ProposerThread class, name tentative. Also added a temporary message types enum to be used in the proposer thread to determine how to handle incoming messages.

Note to self, we will want to reconsider the whole state machine idea because we may not always want to handle recieving a message the same way.