Timeout changes for Proposer Role
Updated our Proposer to use socket.setSoTimeout when waiting for replies in phase one and phase two.
Combined the waitAck and waitPromise functions into a waitForMajority function.
Added explicit units to our Constants.TIMEOUT variable name (ms).
Added a PROMISE_NACK message type. This is a bit contentious of me, but I believe that we can simplify the design of the proposer by explicitly stating what phase the nack was from. This is because we don't want to deal with nacks from phase one when we are considering phase two.